+++ module/actor/actor-sheet.js
@@ -338,12 +338,12 @@ export class ArM5eActorSheet extends ActorSheet {
       }
       context.isDead = this.actor.system.wounds.dead.length > 0;
       context.system.isMagus = this.actor._isMagus();
-      if (context.system.covenant) {
-        if (context.system.covenant.linked) {
-          this.actor.apps[context.system.covenant.document.sheet.appId] =
-            context.system.covenant.document.sheet;
-        }
-      }
+      // if (context.system.covenant) {
+      //   if (context.system.covenant.linked) {
+      //     this.actor.apps[context.system.covenant.document.sheet.appId] =
+      //       context.system.covenant.document.sheet;
+      //   }
+      // }
 
       if (
         context.system?.charType?.value == "magusNPC" ||
@@ -364,12 +364,12 @@ export class ArM5eActorSheet extends ActorSheet {
         context.artsIcons = game.settings.get("arm5e", "artsIcons");
 
         // check whether the character is linked to an existing lab
-        if (context.system.sanctum) {
-          if (context.system.sanctum.linked) {
-            this.actor.apps[context.system.sanctum.document.sheet.appId] =
-              context.system.sanctum.document.sheet;
-          }
-        }
+        // if (context.system.sanctum) {
+        //   if (context.system.sanctum.linked) {
+        //     this.actor.apps[context.system.sanctum.document.sheet.appId] =
+        //       context.system.sanctum.document.sheet;
+        //   }
+        // }
 
         // casting total modifiers
 
@@ -761,6 +761,31 @@ export class ArM5eActorSheet extends ActorSheet {
     return context;
   }
 
+  async _render(force, options = {}) {
+    // Parent class rendering workflow
+    await super._render(force, options);
+
+    // Register the active Application with the referenced Documents
+
+    if (this.actor.system.covenant) {
+      if (this.actor.system.covenant.linked) {
+        this.actor.system.covenant.document.apps[this.appId] = this;
+      }
+    }
+
+    if (this.actor.system.sanctum) {
+      if (this.actor.system.sanctum.linked) {
+        this.actor.system.sanctum.document.apps[this.appId] = this;
+      }
+    }
+
+    if (this.actor.system.owner) {
+      if (this.actor.system.owner.linked) {
+        this.actor.system.owner.document.apps[this.appId] = this;
+      }
+    }
+  }
+
   /**
    * Organize and classify Items for Character sheets.
    *
@@ -773,7 +798,8 @@ export class ArM5eActorSheet extends ActorSheet {
       actorData.actor.type == "player" ||
       actorData.actor.type == "npc" ||
       actorData.actor.type == "laboratory" ||
-      actorData.actor.type == "covenant"
+      actorData.actor.type == "covenant" ||
+      actorData.actor.type == "beast"
     ) {
       for (let virtue of actorData.system.virtues) {
         if (virtue.effects.size > 0) {
@@ -934,7 +960,8 @@ export class ArM5eActorSheet extends ActorSheet {
       let updateArray = [];
       // if the actor was linked, remove listener
       if (this.actor.system.covenant.linked) {
-        delete this.actor.apps[this.actor.system.covenant.document.sheet.appId];
+        delete this.actor.apps[this.actor.system.covenant.document.sheet?.appId];
+        delete this.actor.system.covenant.document.apps[this.appId];
         await this.actor.system.covenant.document.sheet._unbindActor(this.actor);
       }
 
@@ -958,7 +985,8 @@ export class ArM5eActorSheet extends ActorSheet {
       let updateArray = [];
       // if the actor was linked, remove listener
       if (this.actor.system.sanctum.linked) {
-        delete this.actor.apps[this.actor.system.sanctum.document.sheet.appId];
+        delete this.actor.apps[this.actor.system.sanctum.document.sheet?.appId];
+        delete this.actor.system.sanctum.document.apps[this.appId];
         updateArray.push(await this.actor.system.sanctum.document.sheet._unbindActor(this.actor));
       }
       let updateData = { "system.sanctum.value": val };
@@ -1850,14 +1878,16 @@ export class ArM5eActorSheet extends ActorSheet {
 
     if (droppedActor.type === "covenant") {
       if (this.actor.system.covenant.linked) {
-        delete this.actor.apps[this.actor.system.owner.document.sheet.appId];
-        updateArray.push(await this.actor.system.owner.document.sheet._unbindActor(this.actor));
+        delete this.actor.apps[this.actor.system.covenant.document.sheet?.appId];
+        delete this.actor.system.covenant.document.apps[this.appId];
+        updateArray.push(await this.actor.system.covenant.document.sheet._unbindActor(this.actor));
         await droppedActor.sheet._bindActor(this.actor);
       }
     } else if (droppedActor.type === "laboratory") {
       if (this.actor.system.sanctum.linked) {
-        delete this.actor.apps[this.actor.system.owner.document.sheet.appId];
-        updateArray.push(await this.actor.system.owner.document.sheet._unbindActor(this.actor));
+        delete this.actor.apps[this.actor.system.owner.sanctum.sheet?.appId];
+        delete this.actor.system.sanctum.document.apps[this.appId];
+        updateArray.push(await this.actor.system.sanctum.document.sheet._unbindActor(this.actor));
       }
       updateArray.push(await droppedActor.sheet._bindActor(this.actor));
     }