+++ module/arm5e.js
@@ -29,7 +29,7 @@ import { log } from "./tools.js";
 
 import { registerSettings } from "./settings.js";
 import { registerTestSuites } from "./tests/tests.js";
-import { StressDie, StressDieInternal } from "./helpers/stressdie.js";
+import { ArsRoll, StressDie, StressDieInternal } from "./helpers/stressdie.js";
 import { UserguideTour } from "./tours/userguide-tour.js";
 
 import {
@@ -118,33 +118,37 @@ Hooks.once("init", async function () {
   // Experimental
   CONFIG.Dice.types.push(StressDie);
   CONFIG.Dice.types.push(StressDieInternal);
+  CONFIG.Dice.ArsRoll = ArsRoll;
   // CONFIG.Dice.types.push(StressDieNoBotchInternal);
   CONFIG.Dice.terms[StressDie.DENOMINATION] = StressDie;
   CONFIG.Dice.terms[StressDieInternal.DENOMINATION] = StressDieInternal;
-  // CONFIG.Dice.terms[StressDieNoBotchInternal.DENOMINATION] = StressDieNoBotchInternal;
-  // instrumenting roll for testing
-  Roll.prototype.botches = 0;
-  Roll.prototype.diviser = 1;
-  Roll.prototype.multiplier = 1;
-  Roll.prototype.offset = 0;
-  Roll.prototype.modifier = function () {
-    if (!this.result) {
-      return 0;
-    }
-    if (this.botches > 0) {
-      return 0;
-    }
-    if (this.dice.length != 1) {
-      log(false, "ERROR: wrong number of dice");
-      return 0;
-    }
+  CONFIG.Dice.rolls[0] = ArsRoll;
+  //CONFIG.Dice.rolls.push(ArsRoll);
 
-    log(
-      false,
-      `DBG: Roll total ${this.total} * ${this.diviser} - (${this.dice[0].total} * ${this.multiplier}) `
-    );
-    return this.total * this.diviser - this.dice[0].total * this.multiplier;
-  };
+  // // CONFIG.Dice.terms[StressDieNoBotchInternal.DENOMINATION] = StressDieNoBotchInternal;
+  // // instrumenting roll for testing
+  // Roll.prototype.botches = 0;
+  // Roll.prototype.diviser = 1;
+  // Roll.prototype.multiplier = 1;
+  // Roll.prototype.offset = 0;
+  // Roll.prototype.modifier = function () {
+  //   if (!this.result) {
+  //     return 0;
+  //   }
+  //   if (this.botches > 0) {
+  //     return 0;
+  //   }
+  //   if (this.dice.length != 1) {
+  //     log(false, "ERROR: wrong number of dice");
+  //     return 0;
+  //   }
+
+  //   log(
+  //     false,
+  //     `DBG: Roll total ${this.total} * ${this.diviser} - (${this.dice[0].total} * ${this.multiplier}) `
+  //   );
+  //   return this.total * this.diviser - this.dice[0].total * this.multiplier;
+  // };
 
   // UI customization
   CONFIG.ui.actors = ArM5eActorsDirectory;
@@ -508,6 +512,10 @@ Hooks.on("renderDialog", (dialog, html) => {
 
 Hooks.on("renderChatMessage", (message, html, data) =>
   Arm5eChatMessage.addChatListeners(message, html, data)
+);
+
+Hooks.on("createChatMessage", (message, html, data) =>
+  Arm5eChatMessage.enrichChatMessage(message, html, data)
 );
 
 // On Apply an ActiveEffect that uses a CUSTOM application mode.