+++ ReleaseNotes.md
@@ -1,3 +1,21 @@
+## 2.3.0.11, Quendalon, raised by the Fae BF4
+
+### Bug fixes
+
+- Fix requisites editing on roll
+- Applied activities can no longer have their type changed.
+- Fixed bug in sanatorium V12 when changing the data.
+- Virtues and flaws are working again on beast sheet
+- Vis source attributes are visible again
+- Quick dialogs (Vitals, Combat, Magic) are now updating when the linked character changes (V12 only)
+- [V12] Changed actor links mechanics since appId of a sheet is only defined on render (instead of constructor).
+- Covenant laborers were wrongly considered servants
+- Covenant teamsters were not considered workers.
+- It is possible again to go to previous topic of a book in scriptorium
+- Missing spell warning is now properly formated in scriptorium
+- Multitopic book reading was not working after a change to the dialog (reader, date, ...)
+- There is no longer an error when you try to read an ability book you are too skilled for.
+
 ## 2.3.0.10, Quendalon, raised by the Fae BF3
 
 ### Bug fixes